Add api to delete logical switch and bindings
Also set BigInteger instead of Integer in the Physical_Port and
Logical_Switch sets, to avoid cast exceptions with Integer despite
declared types.

Adding also a small test class useful for manual testing.

+++ ovsdb/src/main/java/org/opendaylight/ovsdb/lib/table/vtep/Logical_Switch.java
@@ -1,5 +1,7 @@
 package org.opendaylight.ovsdb.lib.table.vtep;
 
+import java.math.BigInteger;
+
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.opendaylight.ovsdb.lib.notation.OvsDBSet;
 import org.opendaylight.ovsdb.lib.table.internal.Table;
@@ -18,7 +20,7 @@ public class Logical_Switch extends Table<Logical_Switch> {
     private String description;
     private String name;
     // It's a plain int in the documentation but the emulator has an array
-    private OvsDBSet<Integer> tunnel_key;
+    private OvsDBSet<BigInteger> tunnel_key;
 
     @Override
     @JsonIgnore
@@ -42,8 +44,8 @@ public class Logical_Switch extends Table<Logical_Switch> {
 
     public void setName(String name) { this.name = name; }
 
-    public OvsDBSet<Integer> getTunnel_key() { return tunnel_key; }
+    public OvsDBSet<BigInteger> getTunnel_key() { return tunnel_key; }
 
-    public void setTunnel_key(OvsDBSet<Integer> key) { this.tunnel_key = key; }
+    public void setTunnel_key(OvsDBSet<BigInteger> key) { this.tunnel_key = key; }
 
 }
